@@ -4,7 +4,7 @@ version: 3.2.0
 
 environment:
   sdk: ">=3.4.3 <4.0.0"
-  flutter: ">=3.24.5"
+  flutter: ">=3.27.2"
 
 dependencies:
   flutter:
@@ -14,12 +14,12 @@ dependencies:
   fbutton_nullsafety: ^5.0.0
   fsuper_nullsafety: ^3.0.3
   common_utils: ^2.1.0
-  extended_image: ^9.0.7
+  extended_image: ^9.1.0
   waterfall_flow: ^3.1.0
   pull_to_refresh_notification: ^3.1.0
-  connectivity_plus: ^6.1.0
+  connectivity_plus: ^6.1.2
   shimmer: ^3.0.0
-  flutter_svg: ^2.0.16
+  flutter_svg: ^2.0.17
   html: ^0.15.5
   after_layout: ^1.2.0
   palette_generator: ^0.3.3+5
@@ -27,10 +27,10 @@ dependencies:
   encrypt: ^5.0.3
   extended_nested_scroll_view: ^6.2.1
   fcontrol_nullsafety: ^2.0.0
-  get_it: ^8.0.2
+  get_it: ^8.0.3
   logger: ^2.5.0
   bubble_tab_indicator: ^0.1.6
-  flutter_markdown: ^0.7.4+3
+  flutter_markdown: ^0.7.5
   package_info: ^2.0.2
   fbroadcast_nullsafety: any
   show_up_animation: ^2.0.0
@@ -41,40 +41,38 @@ dependencies:
   flutter_smart_dialog: ^4.9.8+5
   hive_flutter: ^1.1.0
   flutter_widget_from_html_core: ^0.15.2
-  flex_color_scheme: ^8.0.2
+  flex_color_scheme: ^8.1.0
   freezed_annotation: ^2.4.4
   json_annotation: ^4.9.0
   loading_animation_widget: ^1.3.0
   keyboard_actions: ^4.2.0
-  fast_immutable_collections: ^11.0.0
-  flutter_slidable:
-    git:
-      url: "https://github.com/letsar/flutter_slidable"
-  loading_more_list_fast: ^7.1.8
-  loading_more_list_library_fast: ^4.1.6
-  dd_js_util: ^6.5.2
-  dd_check_plugin: ^3.3.0
+  fast_immutable_collections: ^11.0.3
+  flutter_slidable: any
+  loading_more_list_fast: ^7.1.9
+  loading_more_list_library_fast: ^4.1.7
+  dd_js_util: ^6.5.6
+  dd_check_plugin: ^3.3.1
   permission_handler: ^11.3.1
-  go_router: ^14.6.2
+  go_router: ^14.6.3
   responsive_builder: ^0.7.1
   image_picker: ^1.1.2
-  shared_preferences: ^2.3.3
+  shared_preferences: ^2.3.5
   webview_flutter: ^4.10.0
   async: ^2.12.0
-  share_plus: ^10.1.2
+  share_plus: ^10.1.4
   blurrycontainer: ^2.1.0
   scoped_model: ^2.0.0
   animate_do: ^3.3.4
 
 
 dev_dependencies:
-  build_runner: ^2.4.13
-  freezed: ^2.5.7
-  json_serializable: ^6.9.0
-  hive_generator_plus: ^3.0.5
+  build_runner: ^2.4.14
+  freezed: ^2.5.8
+  json_serializable: ^6.9.3
+  hive_generator_plus: ^4.0.0
 
 dependency_overrides:
-  watcher: ^1.1.0
+  watcher: ^1.1.1
   flutter_lints: ^5.0.0
   web: ^1.1.0
   meta: ^1.16.0
