@@ -5,12 +5,15 @@ class BottomCategoryTabs extends ConsumerWidget implements PreferredSizeWidget {
   final ValueChanged<int>? onTap;
   final int? initIndex;
 
-  const BottomCategoryTabs({this.insets, this.onTap, this.initIndex, super.key});
+  const BottomCategoryTabs(
+      {this.insets, this.onTap, this.initIndex, super.key});
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
     var myInsets = insets ?? [];
-    final categoryWidgets = ref.watch(categoryRiverpod.select((value) => value.categorys)).map((element) => Tab(text: element.cname));
+    final categoryWidgets = ref
+        .watch(categoryRiverpod.select((value) => value.categorys))
+        .map((element) => Tab(text: element.cname));
     return Container(
       alignment: Alignment.centerLeft,
       child: DefaultTabController(
@@ -24,7 +27,7 @@ class BottomCategoryTabs extends ConsumerWidget implements PreferredSizeWidget {
             indicatorColor: Colors.pink,
             tabBarIndicatorSize: TabBarIndicatorSize.tab,
           ),
-          unselectedLabelColor: Colors.black.withOpacity(.67),
+          unselectedLabelColor: Colors.black.withValues(alpha: .67),
           labelColor: Colors.white,
           onTap: onTap,
         ),
