@@ -26,7 +26,11 @@ class OrderAddState extends State<OrderAddIndexPage> {
         child: Column(
           children: <Widget>[
             buildPlaceHolderSizedBox(),
-            Center(child: SvgPicture.asset('assets/svg/undraw_shopping_app_flsj.svg', height: svgSize, width: svgSize)),
+            Center(
+                child: SvgPicture.asset(
+                    'assets/svg/undraw_shopping_app_flsj.svg',
+                    height: svgSize,
+                    width: svgSize)),
             buildPlaceHolderSizedBox(),
             Container(
               margin: const EdgeInsets.symmetric(horizontal: 20),
@@ -52,14 +56,15 @@ class OrderAddState extends State<OrderAddIndexPage> {
               },
               clickEffect: true,
               padding: EdgeInsets.zero,
-              highlightColor: const Color(0xffE65100).withOpacity(0.20),
-              hoverColor: Colors.redAccent.withOpacity(0.16),
+              highlightColor: const Color(0xffE65100).withValues(alpha: 0.20),
+              hoverColor: Colors.redAccent.withValues(alpha: 0.16),
             ),
             buildPlaceHolderSizedBox(),
             FSuper(
               lightOrientation: FLightOrientation.LeftBottom,
               width: double.infinity,
-              padding: const EdgeInsets.fromLTRB((16.0 + 25.0 + 12), 8, (0.0 + 8.0), 8),
+              padding: const EdgeInsets.fromLTRB(
+                  (16.0 + 25.0 + 12), 8, (0.0 + 8.0), 8),
               margin: const EdgeInsets.fromLTRB(10, 10, 10, 0),
               backgroundColor: const Color(0xfffff0e7),
               strokeColor: const Color(0xfffee0cd),
@@ -67,7 +72,11 @@ class OrderAddState extends State<OrderAddIndexPage> {
               text: '注意事项',
               textAlignment: Alignment.centerLeft,
               textAlign: TextAlign.left,
-              spans: const [TextSpan(text: '\n只有通过本站链接购买的订单才能审核通过并获得奖励,否则绑定失败.(多次绑定失败将封号处理)', style: TextStyle(color: Colors.black26))],
+              spans: const [
+                TextSpan(
+                    text: '\n只有通过本站链接购买的订单才能审核通过并获得奖励,否则绑定失败.(多次绑定失败将封号处理)',
+                    style: TextStyle(color: Colors.black26))
+              ],
               child1: Transform.rotate(
                 angle: pi,
                 child: const Icon(
